\docType{data}
\name{bug}
\alias{bug}
\title{Retorna um data.frame com todos os processos retornados pela query.}
\format{\preformatted{ NULL
}}
\usage{
bug
}
\description{
Cuidado! Os parâmetros que a função recebe (livre, classes,
etc) precisam ser exatamente iguais às informações que o
TJSP precisa para retornar as informações.
}
\keyword{datasets}

