\name{crawler_cjpg}
\alias{crawler_cjpg}
\title{Retorna um data.frame com todos os processos retornados pela query.}
\usage{
crawler_cjpg(livre = "", classes = "", assuntos = "", magistrados = "",
  datas = "", varas = "")
}
\description{
Cuidado! Os parâmetros que a função recebe (livre, classes,
etc) precisam ser exatamente iguais às informações que o
TJSP precisa para retornar as informações.
}

